All functions and types

ConformalPrediction.ConformalModels.available_models

ConformalPrediction.ConformalModels.ConformalModel

ConformalPrediction.ConformalModels.InductiveConformalClassifier

ConformalPrediction.ConformalModels.InductiveConformalModel

ConformalPrediction.ConformalModels.InductiveConformalRegressor

ConformalPrediction.ConformalModels.JackknifeRegressor

ConformalPrediction.ConformalModels.NaiveClassifier

ConformalPrediction.ConformalModels.NaiveRegressor

ConformalPrediction.ConformalModels.SimpleInductiveClassifier

ConformalPrediction.ConformalModels.SimpleInductiveRegressor

ConformalPrediction.ConformalModels.TransductiveConformalClassifier

ConformalPrediction.ConformalModels.TransductiveConformalModel

ConformalPrediction.ConformalModels.TransductiveConformalRegressor

ConformalPrediction.ConformalModels.calibrate!

ConformalPrediction.ConformalModels.conformal_model

ConformalPrediction.ConformalModels.empirical_quantile

ConformalPrediction.ConformalModels.predict_region

ConformalPrediction.ConformalModels.predict_region

ConformalPrediction.ConformalModels.predict_region

ConformalPrediction.ConformalModels.predict_region

ConformalPrediction.ConformalModels.predict_region

ConformalPrediction.ConformalModels.score

ConformalPrediction.ConformalModels.score

ConformalPrediction.ConformalModels.score

ConformalPrediction.ConformalModels.score

ConformalPrediction.ConformalModels.score

ConformalPrediction.ConformalModels.score

MLJModelInterface.fit

MLJModelInterface.fit

MLJModelInterface.predict

Exported functions

ConformalPrediction.ConformalModels.available_models — Constant

A container listing all available methods for conformal prediction.

ConformalPrediction.ConformalModels.ConformalModel — Type

An abstract base type for conformal models.

ConformalPrediction.ConformalModels.InductiveConformalModel — Type

An abstract base time of Inductive Conformal Models. These models rely on data splitting. In particular, we partition the training data as $\{1,...,n\}=\mathcal{D}_{\text{train}} \cup \mathcal{D}_{\text{calibration}}$.

ConformalPrediction.ConformalModels.JackknifeRegressor — Type

Constructor for JackknifeRegressor.

ConformalPrediction.ConformalModels.NaiveClassifier — Type

The NaiveClassifier is the simplest approach to Inductive Conformal Classification. Contrary to the NaiveClassifier it computes nonconformity scores using a designated trainibration dataset.

ConformalPrediction.ConformalModels.NaiveRegressor — Type

The NaiveRegressor for conformal prediction is the simplest approach to conformal regression.

ConformalPrediction.ConformalModels.SimpleInductiveClassifier — Type

The SimpleInductiveClassifier is the simplest approach to Inductive Conformal Classification. Contrary to the NaiveClassifier it computes nonconformity scores using a designated calibration dataset.

ConformalPrediction.ConformalModels.SimpleInductiveRegressor — Type

The SimpleInductiveRegressor is the simplest approach to Inductive Conformal Regression. Contrary to the NaiveRegressor it computes nonconformity scores using a designated calibration dataset.

ConformalPrediction.ConformalModels.TransductiveConformalModel — Type

An abstract base time of Transductive Conformal Models. These models do not rely on data splitting. In particular, nonconformity scores are computed using the entire trainign data set $\{1,...,n\}=\mathcal{D}_{\text{train}}$.

ConformalPrediction.ConformalModels.calibrate! — Method

calibrate!(conf_model::InductiveConformalModel, Xcal, ycal)

Calibrates a Inductive Conformal Model using calibration data.

ConformalPrediction.ConformalModels.conformal_model — Method

conformal_model(model::Supervised; method::Union{Nothing, Symbol}=nothing)

A simple wrapper function that turns any modeline{<:Supervised} into a conformal model. It accepts an optional key argument that can be used to specify the desired method for conformal prediction.

ConformalPrediction.ConformalModels.predict_region — Function

predict_region(conf_model::TransductiveConformalRegressor, Xnew, coverage::AbstractFloat=0.95)

Generic method to compute prediction region for given quantile q̂ for Transductive Conformal Regressors.

ConformalPrediction.ConformalModels.predict_region — Function

predict_region(conf_model::InductiveConformalClassifier, Xnew, coverage::AbstractFloat=0.95)

Generic method to compute prediction region for given quantile q̂ for Inductive Conformal Classifiers.

ConformalPrediction.ConformalModels.predict_region — Function

predict_region(conf_model::ConformalModel, Xnew, coverage::AbstractFloat=0.95)

Generic method for generating prediction regions from a calibrated conformal model for a given quantile.

ConformalPrediction.ConformalModels.predict_region — Function

predict_region(conf_model::TransductiveConformalClassifier, Xnew, coverage::AbstractFloat=0.95)

Generic method to compute prediction region for given quantile q̂ for Transductive Conformal Classifiers.

ConformalPrediction.ConformalModels.predict_region — Function

predict_region(conf_model::InductiveConformalRegressor, Xnew, coverage::AbstractFloat=0.95)

Generic method to compute prediction region for given quantile q̂ for Inductive Conformal Regressors.

ConformalPrediction.ConformalModels.score — Method

score(conf_model::ConformalModel, Xcal, ycal)

Generic method for computing non-conformity scores for any conformal model using calibration (inductive) or training (transductive) data.

ConformalPrediction.ConformalModels.score — Method

score(conf_model::JackknifeRegressor, Xtrain, ytrain)

For the JackknifeRegressor prediction intervals are computed as follows,

$\begin{aligned} \hat{C}_{n,\alpha}(X_{n+1}) &= \hat\mu(X_{n+1}) \pm \hat{q}_{n, \alpha}^{+} \{|Y_i - \hat\mu_{-i}(X_i)|\}, \ i \in \mathcal{D}_{\text{train}} \end{aligned}$

where $\hat\mu_{-i}$ denotes the model fitted on training data with $i$th point removed. The jackknife procedure addresses the overfitting issue associated with the NaiveRegressor.

conf_model = conformal_model(model; method=:jackknife)
score(conf_model, X, y)

ConformalPrediction.ConformalModels.score — Method

score(conf_model::NaiveClassifier, Xtrain, ytrain)

For the NaiveClassifier prediction sets are computed as follows:

$\begin{aligned} \hat{C}_{n,\alpha}(X_{n+1}) &= \left\{y: s(X_{n+1},y) \le \hat{q}_{n, \alpha}^{+} \{|Y_i - \hat\mu(X_i) |\} \right\}, \ i \in \mathcal{D}_{\text{train}} \end{aligned}$

The naive approach typically produces prediction regions that undercover due to overfitting.

Examples

conf_model = conformal_model(model; method=:naive)
score(conf_model, X, y)

ConformalPrediction.ConformalModels.score — Method

score(conf_model::NaiveRegressor, Xtrain, ytrain)

For the NaiveRegressor prediction intervals are computed as follows:

$\begin{aligned} \hat{C}_{n,\alpha}(X_{n+1}) &= \hat\mu(X_{n+1}) \pm \hat{q}_{n, \alpha}^{+} \{|Y_i - \hat\mu(X_i)| \}, \ i \in \mathcal{D}_{\text{train}} \end{aligned}$

The naive approach typically produces prediction regions that undercover due to overfitting.

conf_model = conformal_model(model; method=:naive)
score(conf_model, X, y)

ConformalPrediction.ConformalModels.score — Method

score(conf_model::SimpleInductiveClassifier, Xtrain, ytrain)

For the SimpleInductiveClassifier prediction sets are computed as follows,

$\begin{aligned} \hat{C}_{n,\alpha}(X_{n+1}) &= \left\{y: s(X_{n+1},y) \le \hat{q}_{n, \alpha}^{+} \{|Y_i - \hat\mu(X_i) |\} \right\}, \ i \in \mathcal{D}_{\text{calibration}} \end{aligned}$

where $\mathcal{D}_{\text{calibration}}$ denotes the designated calibration data and $\hat\mu$ denotes the model fitted on training data $\mathcal{D}_{\text{train}}$.

Examples

conf_model = conformal_model(model; method=:simple)
score(conf_model, X, y)

ConformalPrediction.ConformalModels.score — Method

score(conf_model::SimpleInductiveRegressor, Xtrain, ytrain)

For the SimpleInductiveRegressor prediction intervals are computed as follows,

$\begin{aligned} \hat{C}_{n,\alpha}(X_{n+1}) &= \hat\mu(X_{n+1}) \pm \hat{q}_{n, \alpha}^{+} \{|Y_i - \hat\mu(X_i)| \}, \ i \in \mathcal{D}_{\text{calibration}} \end{aligned}$

where $\mathcal{D}_{\text{calibration}}$ denotes the designated calibration data and $\hat\mu$ denotes the model fitted on training data $\mathcal{D}_{\text{train}}$.

Examples

conf_model = conformal_model(model; method=:simple)
score(conf_model, X, y)

Internal functions

ConformalPrediction.ConformalModels.InductiveConformalClassifier — Type

A base type for Inductive Conformal Classifiers.

ConformalPrediction.ConformalModels.InductiveConformalRegressor — Type

A base type for Inductive Conformal Regressors.

ConformalPrediction.ConformalModels.TransductiveConformalClassifier — Type

A base type for Transductive Conformal Classifiers.

ConformalPrediction.ConformalModels.TransductiveConformalRegressor — Type

A base type for Transductive Conformal Regressors.

ConformalPrediction.ConformalModels.empirical_quantile — Function

empirical_quantile(conf_model::ConformalModel, coverage::AbstractFloat=0.95)

Computes the empirical quantile q̂ of the calibrated conformal scores for a user chosen coverage rate (1-α).

MLJModelInterface.fit — Method

fit(conf_model::InductiveConformalModel, verbosity, X, y)

Wrapper function to fit the underlying MLJ model. For Inductive Conformal Prediction the underlying model is fitted on the proper training set. The fitresult is assigned to the model instance. Computation of nonconformity scores requires a separate calibration step involving a calibration data set (see calibrate!).

MLJModelInterface.fit — Method

fit(conf_model::TransductiveConformalModel, verbosity, X, y)

Wrapper function to fit the underlying MLJ model and compute nonconformity scores in one single call. This method is only applicable to Transductive Conformal Prediction.

MLJModelInterface.predict — Method

MMI.predict(conf_model::ConformalModel, fitresult, Xnew)

Compulsory generic predict method of MMI. Simply wraps the underlying model and apply generic method to underlying model.