All functions and types

ConformalPrediction.ConformalModels.available_models

ConformalPrediction.ConformalModels.ConformalModel

ConformalPrediction.ConformalModels.InductiveConformalClassifier

ConformalPrediction.ConformalModels.InductiveConformalRegressor

ConformalPrediction.ConformalModels.JackknifeRegressor

ConformalPrediction.ConformalModels.NaiveClassifier

ConformalPrediction.ConformalModels.NaiveRegressor

ConformalPrediction.ConformalModels.SimpleInductiveClassifier

ConformalPrediction.ConformalModels.SimpleInductiveRegressor

ConformalPrediction.ConformalModels.TransductiveConformalClassifier

ConformalPrediction.ConformalModels.TransductiveConformalRegressor

ConformalPrediction.ConformalModels.calibrate!

ConformalPrediction.ConformalModels.conformal_model

ConformalPrediction.ConformalModels.empirical_quantile

ConformalPrediction.ConformalModels.prediction_region

ConformalPrediction.ConformalModels.score

MLJModelInterface.fit

MLJModelInterface.fit

MLJModelInterface.predict

MLJModelInterface.predict

Exported functions

ConformalPrediction.ConformalModels.available_models — Constant

A container listing all available methods for conformal prediction.

ConformalPrediction.ConformalModels.ConformalModel — Type

An abstract base type for conformal models.

ConformalPrediction.ConformalModels.JackknifeRegressor — Type

The Jackknife ...

ConformalPrediction.ConformalModels.NaiveClassifier — Type

The NaiveClassifier is the simplest approach to Inductive Conformal Classification. Contrary to the NaiveClassifier it computes nonconformity scores using a designated trainibration dataset.

ConformalPrediction.ConformalModels.NaiveRegressor — Type

The NaiveRegressor for conformal prediction is the simplest approach to conformal regression. It computes nonconformity scores by simply using the training data.

ConformalPrediction.ConformalModels.SimpleInductiveClassifier — Type

The SimpleInductiveClassifier is the simplest approach to Inductive Conformal Classification. Contrary to the NaiveClassifier it computes nonconformity scores using a designated calibration dataset.

ConformalPrediction.ConformalModels.SimpleInductiveRegressor — Type

The SimpleInductiveRegressor is the simplest approach to Inductive Conformal Regression. Contrary to the NaiveRegressor it computes nonconformity scores using a designated calibration dataset.

ConformalPrediction.ConformalModels.calibrate! — Method

calibrate!(conf_model::InductiveConformalModel, Xcal, ycal)

Calibrates a Inductive Conformal Model using calibration data.

ConformalPrediction.ConformalModels.conformal_model — Method

conformal_model(model::Supervised; method::Union{Nothing, Symbol}=nothing)

A simple wrapper function that turns any modeline{<:Supervised} into a conformal model. It accepts an optional key argument that can be used to specify the desired method for conformal prediction.

ConformalPrediction.ConformalModels.empirical_quantile — Function

empirical_quantile(conf_model::ConformalModel, coverage::AbstractFloat=0.95)

Computes the empirical quantile q̂ of the calibrated conformal scores for a user chosen coverage rate (1-α).

ConformalPrediction.ConformalModels.prediction_region — Method

prediction_region(conf_model::ConformalModel, Xnew, q̂::Real)

Generic method for generating prediction regions from a calibrated conformal model for a given quantile.

ConformalPrediction.ConformalModels.score — Method

score(conf_model::ConformalModel, Xcal, ycal)

Generic method for computing non-conformity scores for any conformal model using calibration (inductive) or training (transductive) data.

MLJModelInterface.fit — Method

fit(conf_model::InductiveConformalModel, verbosity, X, y)

Wrapper function to fit the underlying MLJ model. For Inductive Conformal Prediction the underlying model is fitted on the proper training set. The fitresult is assigned to the model instance. Computation of nonconformity scores requires a separate calibration step involving a calibration data set (see calibrate!).

MLJModelInterface.fit — Method

fit(conf_model::TransductiveConformalModel, verbosity, X, y)

Wrapper function to fit the underlying MLJ model and compute nonconformity scores in one single call. This method is only applicable to Transductive Conformal Prediction.

Internal functions

ConformalPrediction.ConformalModels.InductiveConformalClassifier — Type

A base type for Inductive Conformal Classifiers.

ConformalPrediction.ConformalModels.InductiveConformalRegressor — Type

A base type for Inductive Conformal Regressors.

ConformalPrediction.ConformalModels.TransductiveConformalClassifier — Type

A base type for Transductive Conformal Classifiers.

ConformalPrediction.ConformalModels.TransductiveConformalRegressor — Type

A base type for Transductive Conformal Regressors.

MLJModelInterface.predict — Function

MMI.predict(conf_model::ConformalModel, Xnew, coverage::AbstractFloat=0.95)

Computes the conformal prediction for any calibrated conformal model and new data Xnew. The default coverage ratio (1-α) is set to 95%.

MLJModelInterface.predict — Method

MMI.predict(conf_model::ConformalModel, fitresult, Xnew)

Compulsory generic predict method of MMI. Simply wraps the underlying model and apply generic method to underlying model.