All functions and types

ConformalPrediction.ConformalModels.available_models

ConformalPrediction.ConformalModels.AdaptiveInductiveClassifier

ConformalPrediction.ConformalModels.CVMinMaxRegressor

ConformalPrediction.ConformalModels.CVPlusRegressor

ConformalPrediction.ConformalModels.ConformalModel

ConformalPrediction.ConformalModels.InductiveConformalClassifier

ConformalPrediction.ConformalModels.InductiveConformalModel

ConformalPrediction.ConformalModels.InductiveConformalRegressor

ConformalPrediction.ConformalModels.JackknifeMinMaxRegressor

ConformalPrediction.ConformalModels.JackknifePlusRegressor

ConformalPrediction.ConformalModels.JackknifeRegressor

ConformalPrediction.ConformalModels.NaiveClassifier

ConformalPrediction.ConformalModels.NaiveRegressor

ConformalPrediction.ConformalModels.SimpleInductiveClassifier

ConformalPrediction.ConformalModels.SimpleInductiveRegressor

ConformalPrediction.ConformalModels.TransductiveConformalClassifier

ConformalPrediction.ConformalModels.TransductiveConformalModel

ConformalPrediction.ConformalModels.TransductiveConformalRegressor

ConformalPrediction.ConformalModels.conformal_model

MLJModelInterface.fit

MLJModelInterface.fit

MLJModelInterface.fit

MLJModelInterface.fit

MLJModelInterface.fit

MLJModelInterface.fit

MLJModelInterface.fit

MLJModelInterface.fit

MLJModelInterface.fit

MLJModelInterface.fit

MLJModelInterface.fit

MLJModelInterface.predict

MLJModelInterface.predict

MLJModelInterface.predict

MLJModelInterface.predict

MLJModelInterface.predict

MLJModelInterface.predict

MLJModelInterface.predict

MLJModelInterface.predict

MLJModelInterface.predict

MLJModelInterface.predict

MLJModelInterface.predict

Exported functions

ConformalPrediction.ConformalModels.available_models — Constant

A container listing all available methods for conformal prediction.

ConformalPrediction.ConformalModels.AdaptiveInductiveClassifier — Type

The AdaptiveInductiveClassifier is the simplest approach to Inductive Conformal Classification. Contrary to the NaiveClassifier it computes nonconformity scores using a designated calibration dataset.

ConformalPrediction.ConformalModels.CVMinMaxRegressor — Type

Constructor for CVMinMaxRegressor.

ConformalPrediction.ConformalModels.CVPlusRegressor — Type

Constructor for CVPlusRegressor.

ConformalPrediction.ConformalModels.ConformalModel — Type

An abstract base type for conformal models.

ConformalPrediction.ConformalModels.InductiveConformalClassifier — Type

A base type for Inductive Conformal Classifiers.

ConformalPrediction.ConformalModels.InductiveConformalModel — Type

An abstract base time of Inductive Conformal Models. These models rely on data splitting. In particular, we partition the training data as $\{1,...,n\}=\mathcal{D}_{\text{train}} \cup \mathcal{D}_{\text{calibration}}$.

ConformalPrediction.ConformalModels.InductiveConformalRegressor — Type

A base type for Inductive Conformal Regressors.

ConformalPrediction.ConformalModels.JackknifeMinMaxRegressor — Type

Constructor for JackknifeMinMaxRegressor.

ConformalPrediction.ConformalModels.JackknifePlusRegressor — Type

Constructor for JackknifePlusRegressor.

ConformalPrediction.ConformalModels.JackknifeRegressor — Type

Constructor for JackknifeRegressor.

ConformalPrediction.ConformalModels.NaiveClassifier — Type

The NaiveClassifier is the simplest approach to Inductive Conformal Classification. Contrary to the NaiveClassifier it computes nonconformity scores using a designated trainibration dataset.

ConformalPrediction.ConformalModels.NaiveRegressor — Type

The NaiveRegressor for conformal prediction is the simplest approach to conformal regression.

ConformalPrediction.ConformalModels.SimpleInductiveClassifier — Type

The SimpleInductiveClassifier is the simplest approach to Inductive Conformal Classification. Contrary to the NaiveClassifier it computes nonconformity scores using a designated calibration dataset.

ConformalPrediction.ConformalModels.SimpleInductiveRegressor — Type

The SimpleInductiveRegressor is the simplest approach to Inductive Conformal Regression. Contrary to the NaiveRegressor it computes nonconformity scores using a designated calibration dataset.

ConformalPrediction.ConformalModels.TransductiveConformalClassifier — Type

A base type for Transductive Conformal Classifiers.

ConformalPrediction.ConformalModels.TransductiveConformalModel — Type

An abstract base time of Transductive Conformal Models. These models do not rely on data splitting. In particular, nonconformity scores are computed using the entire trainign data set $\{1,...,n\}=\mathcal{D}_{\text{train}}$.

ConformalPrediction.ConformalModels.TransductiveConformalRegressor — Type

A base type for Transductive Conformal Regressors.

ConformalPrediction.ConformalModels.conformal_model — Method

conformal_model(model::Supervised; method::Union{Nothing, Symbol}=nothing, kwargs...)

A simple wrapper function that turns a model::Supervised into a conformal model. It accepts an optional key argument that can be used to specify the desired method for conformal prediction as well as additinal kwargs... specific to the method.

Internal functions

MLJModelInterface.fit — Method

MMI.fit(conf_model::ConformalModel, verbosity, X, y)

Generic fit method used to train conformal models. If no specific fit method is dispatched for conf_model::ConformalModel, calling fit defaults to simply fitting the underling atomic model.

MLJModelInterface.fit — Method

MMI.fit(conf_model::AdaptiveInductiveClassifier, verbosity, X, y)

Wrapper function to fit the underlying MLJ model.

MLJModelInterface.fit — Method

MMI.fit(conf_model::CVMinMaxRegressor, verbosity, X, y)

Wrapper function to fit the underlying MLJ model.

MLJModelInterface.fit — Method

MMI.fit(conf_model::CVPlusRegressor, verbosity, X, y)

Wrapper function to fit the underlying MLJ model.

MLJModelInterface.fit — Method

MMI.fit(conf_model::JackknifeMinMaxRegressor, verbosity, X, y)

Wrapper function to fit the underlying MLJ model.

MLJModelInterface.fit — Method

MMI.fit(conf_model::JackknifeRegressor, verbosity, X, y)

Wrapper function to fit the underlying MLJ model.

MLJModelInterface.fit — Method

MMI.fit(conf_model::JackknifeRegressor, verbosity, X, y)

Wrapper function to fit the underlying MLJ model.

MLJModelInterface.fit — Method

MMI.fit(conf_model::NaiveClassifier, verbosity, X, y)

Wrapper function to fit the underlying MLJ model.

MLJModelInterface.fit — Method

MMI.fit(conf_model::NaiveRegressor, verbosity, X, y)

Wrapper function to fit the underlying MLJ model.

MLJModelInterface.fit — Method

MMI.fit(conf_model::SimpleInductiveClassifier, verbosity, X, y)

Wrapper function to fit the underlying MLJ model.

MLJModelInterface.fit — Method

MMI.fit(conf_model::SimpleInductiveRegressor, verbosity, X, y)

Wrapper function to fit the underlying MLJ model.

MLJModelInterface.predict — Method

MMI.predict(conf_model::ConformalModel, fitresult, Xnew)

Generic MMI.predict method used to predict from a conformal model given a fitresult and data Xnew. If no specific predict method is dispatched for conf_model::ConformalModel, calling predict defaults to simply predicting from the underlying atomic model.

MLJModelInterface.predict — Method

MMI.predict(conf_model::AdaptiveInductiveClassifier, fitresult, Xnew)

For the AdaptiveInductiveClassifier nonconformity scores are computed by cumulatively summing the ranked scores of each label in descending order until reaching the true label $y_i$:

$s_i(X_i,y_i) = \sum_{j=1}^k \hat\mu(X_i)_{\pi_j} \ \text{where } \ y_i=\pi_k, i \in \mathcal{D}_{\text{calibration}}$

Prediction sets are then computed as follows,

$\hat{C}_{n,\alpha}(X_{n+1}) = \left\{y: s(X_{n+1},y) \le \hat{q}_{n, \alpha}^{+} \{s_i(X_i,y_i)\} \right\}, i \in \mathcal{D}_{\text{calibration}}$

where $\mathcal{D}_{\text{calibration}}$ denotes the designated calibration data and $\hat\mu$ denotes the model fitted on training data $\mathcal{D}_{\text{train}}$.

MLJModelInterface.predict — Method

MMI.predict(conf_model::CVMinMaxRegressor, fitresult, Xnew)

For the CVMinMaxRegressor prediction intervals are computed as follows,

$\hat{C}_{n,\alpha}(X_{n+1}) = \left[ \min_{i=1,...,n} \hat\mu_{-\mathcal{D}_{k(i)}}(X_{n+1}) - \hat{q}_{n, \alpha}^{+} \{R_i^{\text{CV}} \}, \max_{i=1,...,n} \hat\mu_{-\mathcal{D}_{k(i)}}(X_{n+1}) + \hat{q}_{n, \alpha}^{+} \{ R_i^{\text{CV}}\} \right] , i \in \mathcal{D}_{\text{train}}$

with

$R_i^{\text{CV}}=|Y_i - \hat\mu_{-\mathcal{D}_{k(i)}}(X_i)|$

where $\hat\mu_{-\mathcal{D}_{k(i)}}$ denotes the model fitted on training data with subset $\mathcal{D}_{k(i)}$ that contains the $i$ th point removed.

MLJModelInterface.predict — Method

MMI.predict(conf_model::CVPlusRegressor, fitresult, Xnew)

For the CVPlusRegressor prediction intervals are computed as follows,

$\hat{C}_{n,\alpha}(X_{n+1}) = \left[ \hat{q}_{n, \alpha}^{-} \{\hat\mu_{-\mathcal{D}_{k(i)}}(X_{n+1}) - R_i^{\text{CV}} \}, \hat{q}_{n, \alpha}^{+} \{\hat\mu_{-\mathcal{D}_{k(i)}}(X_{n+1}) + R_i^{\text{CV}}\} \right] , \ i \in \mathcal{D}_{\text{train}}$

with

$R_i^{\text{CV}}=|Y_i - \hat\mu_{-\mathcal{D}_{k(i)}}(X_i)|$

where $\hat\mu_{-\mathcal{D}_{k(i)}}$ denotes the model fitted on training data with subset $\mathcal{D}_{k(i)}$ that contains the $i$ th point removed.

MLJModelInterface.predict — Method

MMI.predict(conf_model::JackknifeMinMaxRegressor, fitresult, Xnew)

For the JackknifeMinMaxRegressor prediction intervals are computed as follows,

$\hat{C}_{n,\alpha}(X_{n+1}) = \left[ \min_{i=1,...,n} \hat\mu_{-i}(X_{n+1}) - \hat{q}_{n, \alpha}^{+} \{R_i^{\text{LOO}} \}, \max_{i=1,...,n} \hat\mu_{-i}(X_{n+1}) + \hat{q}_{n, \alpha}^{+} \{ R_i^{\text{LOO}}\} \right] , i \in \mathcal{D}_{\text{train}}$

with

$R_i^{\text{LOO}}=|Y_i - \hat\mu_{-i}(X_i)|$

where $\hat\mu_{-i}$ denotes the model fitted on training data with $i$th point removed. The jackknife-minmax procedure is more conservative than the JackknifePlusRegressor.

MLJModelInterface.predict — Method

MMI.predict(conf_model::JackknifePlusRegressor, fitresult, Xnew)

For the JackknifePlusRegressor prediction intervals are computed as follows,

$\hat{C}_{n,\alpha}(X_{n+1}) = \left[ \hat{q}_{n, \alpha}^{-} \{\hat\mu_{-i}(X_{n+1}) - R_i^{\text{LOO}} \}, \hat{q}_{n, \alpha}^{+} \{\hat\mu_{-i}(X_{n+1}) + R_i^{\text{LOO}}\} \right] , i \in \mathcal{D}_{\text{train}}$

with

$R_i^{\text{LOO}}=|Y_i - \hat\mu_{-i}(X_i)|$

where $\hat\mu_{-i}$ denotes the model fitted on training data with $i$th point removed. The jackknife$+$ procedure is more stable than the JackknifeRegressor.

MLJModelInterface.predict — Method

MMI.predict(conf_model::JackknifeRegressor, fitresult, Xnew)

For the JackknifeRegressor prediction intervals are computed as follows,

$\hat{C}_{n,\alpha}(X_{n+1}) = \hat\mu(X_{n+1}) \pm \hat{q}_{n, \alpha}^{+} \{|Y_i - \hat\mu_{-i}(X_i)|\}, \ i \in \mathcal{D}_{\text{train}}$

where $\hat\mu_{-i}$ denotes the model fitted on training data with $i$th point removed. The jackknife procedure addresses the overfitting issue associated with the NaiveRegressor.

MLJModelInterface.predict — Method

MMI.predict(conf_model::NaiveClassifier, fitresult, Xnew)

For the NaiveClassifier prediction sets are computed as follows:

$\hat{C}_{n,\alpha}(X_{n+1}) = \left\{y: s(X_{n+1},y) \le \hat{q}_{n, \alpha}^{+} \{1 - \hat\mu(X_i) \} \right\}, \ i \in \mathcal{D}_{\text{train}}$

The naive approach typically produces prediction regions that undercover due to overfitting.

MLJModelInterface.predict — Method

MMI.predict(conf_model::NaiveRegressor, fitresult, Xnew)

For the NaiveRegressor prediction intervals are computed as follows:

$\hat{C}_{n,\alpha}(X_{n+1}) = \hat\mu(X_{n+1}) \pm \hat{q}_{n, \alpha}^{+} \{|Y_i - \hat\mu(X_i)| \}, \ i \in \mathcal{D}_{\text{train}}$

The naive approach typically produces prediction regions that undercover due to overfitting.

MLJModelInterface.predict — Method

MMI.predict(conf_model::SimpleInductiveClassifier, fitresult, Xnew)

For the SimpleInductiveClassifier prediction sets are computed as follows,

$\hat{C}_{n,\alpha}(X_{n+1}) = \left\{y: s(X_{n+1},y) \le \hat{q}_{n, \alpha}^{+} \{1 - \hat\mu(X_i)\} \right\}, \ i \in \mathcal{D}_{\text{calibration}}$

where $\mathcal{D}_{\text{calibration}}$ denotes the designated calibration data and $\hat\mu$ denotes the model fitted on training data $\mathcal{D}_{\text{train}}$.

MLJModelInterface.predict — Method

MMI.predict(conf_model::SimpleInductiveRegressor, fitresult, Xnew)

For the SimpleInductiveRegressor prediction intervals are computed as follows,

$\hat{C}_{n,\alpha}(X_{n+1}) = \hat\mu(X_{n+1}) \pm \hat{q}_{n, \alpha}^{+} \{|Y_i - \hat\mu(X_i)| \}, \ i \in \mathcal{D}_{\text{calibration}}$

where $\mathcal{D}_{\text{calibration}}$ denotes the designated calibration data and $\hat\mu$ denotes the model fitted on training data $\mathcal{D}_{\text{train}}$.