All functions and types

ConformalPrediction.ConformalMachines.available_machines

ConformalPrediction.ConformalMachines.ConformalMachine

ConformalPrediction.ConformalMachines.LABELConformalClassifier

ConformalPrediction.ConformalMachines.NaiveConformalRegressor

ConformalPrediction.ConformalMachines.conformal_machine

ConformalPrediction.ConformalMachines.prediction_region

ConformalPrediction.ConformalMachines.score

ConformalPrediction.calibrate!

MLJModelInterface.predict

Exported functions

ConformalPrediction.calibrate! — Method

calibrate!(conf_mach::ConformalMachine, Xcal, ycal)

Calibrates a conformal machine using calibration data.

MLJModelInterface.predict — Method

predict(conf_mach::ConformalMachine, Xnew; coverage=0.95)

Computes the conformal prediction for any conformal machine and new data Xnew. The default coverage ratio is set to 95%.

ConformalPrediction.ConformalMachines.ConformalMachine — Type

An abstract base type for conformal machines.

ConformalPrediction.ConformalMachines.LABELConformalClassifier — Type

The LABEL method for conformal prediction is the simplest approach to classification.

ConformalPrediction.ConformalMachines.NaiveConformalRegressor — Type

The Naive method for conformal prediction is the simplest approach to regression.

ConformalPrediction.ConformalMachines.conformal_machine — Method

conformal_machine(mach::Machine{<:Supervised}; method::Union{Nothing, Symbol}=nothing)

A simple wrapper function that turns any MLJ.Machine{<:Supervised} into a conformal machine. It accepts an optional key argument that can be used to specify the desired method for conformal prediction.

ConformalPrediction.ConformalMachines.prediction_region — Method

prediction_region(conf_mach::ConformalMachine, Xnew, ϵ::Real)

Generic method for generating prediction regions from a calibrated conformal machine.

ConformalPrediction.ConformalMachines.score — Method

score(conf_mach::ConformalMachine, Xcal, ycal)

Generic method for computing non-conformity scores for any conformal machine using calibration data.

Internal functions

ConformalPrediction.ConformalMachines.available_machines — Constant

A container listing all available methods for conformal prediction.